dto data type changes

+++ src/main/java/com/hcx/hcxprovider/dto/AmountDetailsDTO.java
@@ -4,14 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class AmountDetailsDTO {
-    private Double claimedAmountAfterDiscount;
-    private Double consideredAmountTotal;
-    private Double discountInHospitalBill;
+    private BigDecimal claimedAmountAfterDiscount;
+    private BigDecimal consideredAmountTotal;
+    private BigDecimal discountInHospitalBill;
     private List<ClaimedAmountBreakupDTO> claimedAmountBreakup;
 }
